+++ src/hooks/useFetchPoints.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import axios from "axios";
-import { API_DOMAIN } from "../utils/Constants";
+import { API_DOMAIN } from "../utils/constants";
 import { useAuth } from "../contexts/AuthProvider";
 import { useCart } from "../contexts/CartProvider";
