+++ src/routes/ambassadors/AmbassadorDetail.jsx
@@ -4,7 +4,7 @@ import HeroImage from "../../components/heroImage/HeroImage";
 import Transition from "../../components/transition/Transition";
 import PageTemplate from "../../components/pageTemplate/PageTemplate";
 import useFetch from "../../hooks/useFetch";
-import { API_DOMAIN } from "../../utils/Constants";
+import { API_DOMAIN } from "../../utils/constants";
 import PageGap from "../../components/pageGap/PageGap";
 
 function AmbassadorDetail() {
