+++ src/routes/ambassadorsOverview/AmbassadorsOverview.jsx
@@ -1,6 +1,6 @@
 import Transition from "../../components/transition/Transition";
 import useFetch from "../../hooks/useFetch";
-import { API_DOMAIN } from "../../utils/Constants";
+import { API_DOMAIN } from "../../utils/constants";
 import OverviewPage from "../../components/overviewPage/OverviewPage";
 import PageTemplate from "../../components/pageTemplate/PageTemplate";
