+++ src/routes/home/Home.jsx
@@ -1,7 +1,7 @@
 import Typography from "../../components/typography/Typography";
 import Transition from "../../components/transition/Transition";
 import useFetch from "../../hooks/useFetch";
-import { API_DOMAIN } from "../../utils/Constants";
+import { API_DOMAIN } from "../../utils/constants";
 import PageTemplate from "../../components/pageTemplate/PageTemplate";
 import React, { useState, useEffect } from "react";
 import styles from "./Home.module.css";
