+++ src/routes/login/Login.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
-import { LOGIN_DOMAIN } from "../../utils/Constants";
+import { LOGIN_DOMAIN } from "../../utils/constants";
 import Transition from "../../components/transition/Transition";
 import PageTemplate from "../../components/pageTemplate/PageTemplate";
 import Typography from "../../components/typography/Typography";
