+++ src/routes/projects/ProjectDetail.jsx
@@ -1,6 +1,6 @@
 import { useParams, Link } from "react-router-dom";
 import useFetch from "../../hooks/useFetch";
-import { API_DOMAIN } from "../../utils/Constants";
+import { API_DOMAIN } from "../../utils/constants";
 import Transition from "../../components/transition/Transition";
 import PageTemplate from "../../components/pageTemplate/PageTemplate";
