+++ src/routes/tinkeringOverview/TinkeringOverview.jsx
@@ -1,6 +1,6 @@
 import Transition from "../../components/transition/Transition";
 import useFetch from "../../hooks/useFetch";
-import { API_DOMAIN } from "../../utils/Constants";
+import { API_DOMAIN } from "../../utils/constants";
 import PageTemplate from "../../components/pageTemplate/PageTemplate";
 import { useLenis } from "lenis/react";
 import BackButton from "../../components/BackButton/BackButton";
